Implemented controllers for posts and users
