+++ backend/routes/posts.js
@@ -1,7 +1,8 @@
 const express = require("express");
 const multer = require("multer");
 
-const Post = require('../models/post');
+const PostController = require("../controllers/posts");
+
 const checkAuth = require("../middleware/check-auth");
 
 const router = express.Router();
@@ -33,30 +34,7 @@ router.post(
   "",
   checkAuth,
   multer({storage: storage}).single("image"),
-  (req, res, next) => {
-    const url = req.protocol + '://' + req.get("host");
-    const post = new Post({
-      title: req.body.title,
-      content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename,
-      creator: req.userData.userId
-    });
-    post.save().then(createdPost => {
-      res.status(201).json({
-        message: 'Post added successfully',
-        post: {
-         //get all properties of created post
-         ...createdPost,
-         id: createdPost._id
-        }
-      });
-    })
-    .catch(error => {
-      res.status(500).json({
-        message: "Creating a post failed!"
-      });
-    });
-  }
+  PostController.createPost
 );
 
 // edit existing post
@@ -65,95 +43,16 @@ router.put(
   "/:id",
   checkAuth,
   multer({storage: storage}).single("image"),
-  (req, res, next) => {
-    let imagePath = req.body.imagePath;
-    if (req.file) {
-      const url = req.protocol + '://' + req.get("host");
-      imagePath = url + "/images/" + req.file.filename
-    }
-    const post = new Post({
-      _id: req.body.id,
-      title: req.body.title,
-      content: req.body.content,
-      imagePath: imagePath,
-      creator: req.userData.userId
-    });
-    // updateOne takes arguments of post id and creator id
-    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
-      if(result.nModified > 0) {
-        res.status(200).json({message: "Update successful!"});
-      } else {
-        res.status(401).json({message: "User not authorized to edit"});
-      }
-    })
-    .catch(error => {
-      res.status(500).json({
-        message: "Couldn't update post!"
-      });
-    });
-  }
+  PostController.updatePost
 );
 
 // get all posts
-router.get("", (req, res, next) => {
-  const pageSize = +req.query.pagesize;
-  const currentPage = +req.query.page;
-  const postQuery = Post.find();
-  let fetchedPosts;
-  if (pageSize && currentPage) {
-    postQuery
-      .skip(pageSize * (currentPage - 1))
-      .limit(pageSize);
-  }
-  postQuery
-    .then(documents => {
-      fetchedPosts = documents;
-      return Post.count();
-    })
-    .then(count => {
-      res.status(200).json({
-        message: 'Posts fetched successfully!',
-        posts: fetchedPosts,
-        maxPosts: count
-      });
-    })
-    .catch(error => {
-      res.status(500).json({
-        message: "Fetching posts failed!"
-      });
-    });
-});
+router.get("", PostController.getPosts);
 
 // get single post by id
-router.get("/:id", (req, res, next) => {
-  Post.findById(req.params.id).then(post => {
-    if (post){
-      res.status(200).json(post);
-    } else {
-      res.status(404).json({message: "Post not found!"});
-    }
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Fetching posts failed!"
-    });
-  });
-});
+router.get("/:id", PostController.getPost);
 
 // delete post
-router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
-    if(result.n > 0) {
-      res.status(200).json({message: "Deletion successful!"});
-    } else {
-      res.status(401).json({message: "Not authorized!"});
-    }
-  })
-  .catch(error => {
-    res.status(500).json({
-      message: "Fetching posts failed!"
-    });
-  });
-});
+router.delete("/:id", checkAuth, PostController.deletePost);
 
 module.exports = router;
